+++ leeger/calculator/all_time_calculator/SSLAllTimeCalculator.py
@@ -46,7 +46,7 @@ class SSLAllTimeCalculator(AllTimeCalculator):
             }
         """
 
-        return cls._addAndCombineResults(league, SSLYearCalculator.getTeamScore, validateLeague=False, **kwargs)
+        return cls._addAndCombineResults(league, SSLYearCalculator.getTeamScore, **kwargs)
 
     @classmethod
     @validateLeague
@@ -68,7 +68,7 @@ class SSLAllTimeCalculator(AllTimeCalculator):
             ...
             }
         """
-        return cls._addAndCombineResults(league, SSLYearCalculator.getTeamSuccess, validateLeague=False, **kwargs)
+        return cls._addAndCombineResults(league, SSLYearCalculator.getTeamSuccess, **kwargs)
 
     @classmethod
     @validateLeague
@@ -90,4 +90,4 @@ class SSLAllTimeCalculator(AllTimeCalculator):
             ...
             }
         """
-        return cls._addAndCombineResults(league, SSLYearCalculator.getTeamLuck, validateLeague=False, **kwargs)
+        return cls._addAndCombineResults(league, SSLYearCalculator.getTeamLuck, **kwargs)