+++ leeger/validate/yearValidation.py
@@ -194,7 +194,7 @@ def checkEveryTeamInYearIsInAMatchup(year: Year) -> None:
     Checks that every Team in the year appears in at least 1 matchup.
     """
     from leeger.util.navigator.YearNavigator import YearNavigator
-    teamIds = YearNavigator.getAllTeamIds(year, validateYear=False)
+    teamIds = YearNavigator.getAllTeamIds(year, validate=False)
     for week in year.weeks:
         for matchup in week.matchups:
             try: