+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 - Added way to get League objects in a Python dictionary / JSON via the toJson() method
+- Now allow users to turn off validation on methods that use League objects via the `validate=False` keyword argument
 
 ## [1.0.0]